Random image app
Terms of Reference

Create an HTML page with a gallery. Add a link to generate new images from picsum.photos. Add author's signature. Add styles to the header. Stylize the images. Add color to the link. Stylize author's signature. Replace table with images with flex display. Spread the images across the width. For screens smaller than 360px, make the image width 100%. Create a button inside the link, place it in the center of the row. Make the color change smoothly when hovering over the button.